requirements.txt

streamlit
torch
torchvision
Pillow
matplotlib
scikit-learn
seaborn
